move get_vertexcolor to _matplotlib_backend from html_surface
