+++ nilearn/plotting/surface/html_surface.py
@@ -3,8 +3,6 @@
 import json
 from warnings import warn
 
-import matplotlib as mpl
-import matplotlib.pyplot as plt
 import numpy as np
 
 from nilearn import DEFAULT_DIVERGING_CMAP
@@ -18,11 +16,12 @@ from nilearn.plotting.js_plotting_utils import (
     colorscale,
     get_html_template,
     mesh_to_plotly,
-    to_color_strings,
 )
 from nilearn.plotting.surface._utils import (
+    DEFAULT_ENGINE,
     DEFAULT_HEMI,
     check_surface_plotting_inputs,
+    get_surface_backend,
 )
 from nilearn.surface import (
     PolyMesh,
@@ -40,60 +39,6 @@ from nilearn.surface.surface import (
 
 class SurfaceView(HTMLDocument):  # noqa: D101
     pass
-
-
-def get_vertexcolor(
-    surf_map,
-    cmap,
-    norm,
-    absolute_threshold=None,
-    bg_map=None,
-    bg_on_data=None,
-    darkness=None,
-):
-    """Get the color of the vertices."""
-    if bg_map is None:
-        bg_data = np.ones(len(surf_map)) * 0.5
-        bg_vmin, bg_vmax = 0, 1
-    else:
-        bg_data = np.copy(load_surf_data(bg_map))
-
-    # scale background map if need be
-    bg_vmin, bg_vmax = np.min(bg_data), np.max(bg_data)
-    if bg_vmin < 0 or bg_vmax > 1:
-        bg_norm = mpl.colors.Normalize(vmin=bg_vmin, vmax=bg_vmax)
-        bg_data = bg_norm(bg_data)
-
-    if darkness is not None:
-        bg_data *= darkness
-        warn(
-            (
-                "The `darkness` parameter will be deprecated in release 0.13. "
-                "We recommend setting `darkness` to None"
-            ),
-            DeprecationWarning,
-            stacklevel=find_stack_level(),
-        )
-
-    bg_colors = plt.get_cmap("Greys")(bg_data)
-
-    # select vertices which are filtered out by the threshold
-    if absolute_threshold is None:
-        under_threshold = np.zeros_like(surf_map, dtype=bool)
-    else:
-        under_threshold = np.abs(surf_map) < absolute_threshold
-
-    surf_colors = cmap(norm(surf_map).data)
-    # set transparency of voxels under threshold to 0
-    surf_colors[under_threshold, 3] = 0
-    if bg_on_data:
-        # if need be, set transparency of voxels above threshold to 0.7
-        # so that background map becomes visible
-        surf_colors[~under_threshold, 3] = 0.7
-
-    vertex_colors = cm.mix_colormaps(surf_colors, bg_colors)
-
-    return to_color_strings(vertex_colors)
 
 
 def _one_mesh_info(
@@ -125,7 +70,8 @@ def _one_mesh_info(
         vmin=vmin,
     )
     info = {"inflated_both": mesh_to_plotly(surf_mesh)}
-    info["vertexcolor_both"] = get_vertexcolor(
+    backend = get_surface_backend(DEFAULT_ENGINE)
+    info["vertexcolor_both"] = backend.get_vertexcolor(
         surf_map,
         colors["cmap"],
         colors["norm"],
@@ -244,7 +190,8 @@ def _full_brain_info(
         info[f"pial_{hemi}"] = mesh_to_plotly(mesh[f"pial_{hemi}"])
         info[f"inflated_{hemi}"] = mesh_to_plotly(mesh[f"infl_{hemi}"])
 
-        info[f"vertexcolor_{hemi}"] = get_vertexcolor(
+        backend = get_surface_backend(DEFAULT_ENGINE)
+        info[f"vertexcolor_{hemi}"] = backend.get_vertexcolor(
             surf_map,
             colors["cmap"],
             colors["norm"],
@@ -274,7 +221,8 @@ def _full_brain_info(
                     )
                 )
             )
-    info["vertexcolor_both"] = get_vertexcolor(
+    backend = get_surface_backend(DEFAULT_ENGINE)
+    info["vertexcolor_both"] = backend.get_vertexcolor(
         get_data(surface_maps),
         colors["cmap"],
         colors["norm"],
